refactor(spin): Verantwortlichkeiten in spin.ts aufteilen
- Multiplier-Logik (getMultiplier, setMultiplierSlider, updateMultiplierDisplay,
  initMultiplierSlider) nach multiplier.ts ausgelagert
- resolveWinner() aus spin.ts nach winner.ts verschoben
- Namen-Snapshot (names: string[]) zu SpinConfig hinzugefuegt; spinWheel()
  nimmt Namen als Parameter statt live getNames() aufzurufen
- Imports in main.ts und share-name-list.ts angepasst

Closes #281
