+++ public/script/app/main.ts
@@ -12,9 +12,10 @@ import { nameState } from "../names/name-state.js";
 import { initShareFeature } from "../names/share-name-list.js";
 import { initProfileUI } from "../profile/profiles.js";
 import {
-  initMultiplierSlider, initWheelControls, spinWheel,
-  setSpinOverride, lockSpinButtons, unlockSpinButtons, getMultiplier,
+  initWheelControls, spinWheel,
+  setSpinOverride, lockSpinButtons, unlockSpinButtons,
 } from "../wheel/spin.js";
+import { initMultiplierSlider, getMultiplier } from "../wheel/multiplier.js";
 import { initVolumeSlider } from "../wheel/volume.js";
 import { preloadStaticSounds } from "../wheel/sound.js";
 import { initWinnerModal } from "../wheel/winner.js";
@@ -124,8 +125,9 @@ function clearRoom(): void {
 function handleRoomSpinEvent(lastSpin: number): void {
   if (isHost) return; // host already spun directly from POST response
   lockSpinButtons();
+  const names = getNames();
   const totalSteps = Math.round(lastSpin * getMultiplier());
-  spinWheel(totalSteps, 'right', '');
+  spinWheel(totalSteps, 'right', '', names);
 }
 
 // Host only: POST → spin directly (token guaranteed, no race condition)
@@ -136,7 +138,7 @@ async function handleRoomSpinClick(direction: Direction): Promise<void> {
     const names = getNames();
     const { ranNum, spinToken } = await spinRoom(activeRoomKey, names);
     const totalSteps = Math.round(ranNum * getMultiplier());
-    spinWheel(totalSteps, direction, spinToken);
+    spinWheel(totalSteps, direction, spinToken, names);
   } catch (error) {
     console.error('[ROOM] Spin fehlgeschlagen:', error);
     unlockSpinButtons();
